+++ spark/src/main/scala/geotrellis/spark/io/cog/COGLayerMetadata.scala
@@ -40,6 +40,8 @@ case class COGLayerMetadata[K: SpatialComponent](
   extent: Extent,
   crs: CRS
 ) {
+  val zoomRangesInfosSorted: Vector[(ZoomRange, KeyBounds[K])] = zoomRangeInfos.sortBy(_._1)
+
   def combine(other: COGLayerMetadata[K])(implicit ev: Boundable[K]): COGLayerMetadata[K] = {
     val combinedZoomRangeInfos =
       (zoomRangeInfos ++ other.zoomRangeInfos)
@@ -60,23 +62,23 @@ case class COGLayerMetadata[K: SpatialComponent](
   }
 
   private val maxZooms =
-    zoomRangeInfos.map(_._1.maxZoom).toArray
+    zoomRangesInfosSorted.map(_._1.maxZoom).toArray
 
   def zoomRanges: Vector[ZoomRange] =
-    zoomRangeInfos.map(_._1)
+    zoomRangesInfosSorted.map(_._1)
 
   def zoomRangeFor(zoom: Int): ZoomRange =
     zoomRangeInfoFor(zoom)._1
 
   def zoomRangeInfoFor(zoom: Int): (ZoomRange, KeyBounds[K]) = {
-    val i = java.util.Arrays.binarySearch(maxZooms.sorted, zoom)
+    val i = java.util.Arrays.binarySearch(maxZooms, zoom)
     val idx =
       if(i >= 0) { i }
       else {
         ~i //- 1
       }
 
-    zoomRangeInfos.sortBy(_._1).apply(idx)
+    zoomRangesInfosSorted(idx)
   }
 
 
@@ -346,7 +348,7 @@ object COGLayerMetadata {
 
     COGLayerMetadata(
       cellType,
-      zoomRanges.toVector,
+      zoomRanges.toVector.sortBy(_._1),
       layoutScheme,
       extent,
       crs
@@ -358,7 +360,7 @@ object COGLayerMetadata {
       def write(metadata: COGLayerMetadata[K]) =
         JsObject(
           "cellType" -> metadata.cellType.toJson,
-          "zoomRangesInfos" -> metadata.zoomRangeInfos.sortBy(_._1).toJson,
+          "zoomRangesInfos" -> metadata.zoomRangesInfosSorted.toJson,
           "layoutScheme" -> metadata.layoutScheme.toJson,
           "extent" -> metadata.extent.toJson,
           "crs" -> metadata.crs.toJson
@@ -369,13 +371,13 @@ object COGLayerMetadata {
           case Seq(cellType, JsArray(zoomRanges), layoutScheme, extent, crs) =>
             COGLayerMetadata(
               cellType.convertTo[CellType],
-              zoomRanges.map(_.convertTo[(ZoomRange, KeyBounds[K])]).sortBy(_._1).toVector,
+              zoomRanges.map(_.convertTo[(ZoomRange, KeyBounds[K])]).sortBy(_._1),
               layoutScheme.convertTo[ZoomedLayoutScheme],
               extent.convertTo[Extent],
               crs.convertTo[CRS]
             )
           case v =>
-            throw new DeserializationException(s"COGLayerMetadata expected, got $v")
+            throw DeserializationException(s"COGLayerMetadata expected, got $v")
         }
     }
 }